updates to conversations +import of variables, export of bglobals +support for custom strings and direct string look up using a full reference +added supporting values for imported functions such as game_mins, game_time, game_days, weapon_drawn etc

+++ src/conversation/bglobal.cs
@@ -8,8 +8,8 @@ namespace Underworld
     {
         public struct BablGlobal
         {
-            public int ConversationNo;
-            public int Size;
+            public short ConversationNo;
+            public short Size;
             public short[] Globals;
         };
 
@@ -30,8 +30,8 @@ namespace Underworld
                     bGlobals = new BablGlobal[NoOfSlots + 1];
                     for (int i = 0; i <= NoOfSlots; i++)
                     {
-                        bGlobals[i].ConversationNo = (int)Loader.getAt(bglob_data, add_ptr, 16);
-                        bGlobals[i].Size = (int)Loader.getAt(bglob_data, add_ptr + 2, 16);
+                        bGlobals[i].ConversationNo = (short)Loader.getAt(bglob_data, add_ptr, 16);
+                        bGlobals[i].Size = (short)Loader.getAt(bglob_data, add_ptr + 2, 16);
                         bGlobals[i].Globals = new short[bGlobals[i].Size];
                         add_ptr += 4;
                     }
@@ -52,8 +52,8 @@ namespace Underworld
                     for (int i = 0; i < NoOfSlots; i++)
                     {
 
-                        bGlobals[i].ConversationNo = (int)Loader.getAt(bglob_data, add_ptr, 16);
-                        bGlobals[i].Size = (int)Loader.getAt(bglob_data, add_ptr + 2, 16);
+                        bGlobals[i].ConversationNo = (short)Loader.getAt(bglob_data, add_ptr, 16);
+                        bGlobals[i].Size = (short)Loader.getAt(bglob_data, add_ptr + 2, 16);
                         bGlobals[i].Globals = new short[bGlobals[i].Size];
                         add_ptr += 4;
                         for (int g = 0; g < bGlobals[i].Size; g++)
